feat: added typing more rigid
