+++ examples/create-react-app-refresh-token/src/refreshApi.js
@@ -46,7 +46,7 @@ const refreshApi = createRefresh({
       return {
         isSuccess: false
       } 
-    }    
+    }
   }
 })
