+++ packages/react-auth-kit/src/types.ts
@@ -1,114 +1,189 @@
 /**
- *
- * @author Arkadip Bhattacharya <hi@arkadip.dev>
- * @fileoverview General types used in the library
- * @copyright Arkadip Bhattacharya 2020
- *
+ * Parameters for SignIn Action
  */
-
 export interface SignInActionPayload<T> {
+  /**
+   * Authentication object
+   */
   auth: {
+    
+    /**
+     * JWT access token
+     */
     token: string,
+    
+    /**
+     * `type` or `prefix` of the token, to be used in the Network Request
+     * 
+     * @example 
+     * Bearer
+     */
     type?: string
   },
+  
+  /**
+   * JWT refresh token.
+   * 
+   * Make sure, you are using refresh token and configured the `createRefresh`
+   * before adding the token
+   */
   refresh?: string | null,
+  
+  /**
+   * User information
+   */
   userState?: T,
 }
 
-export interface RefreshTokenActionPayload<T> {
-  newAuthToken?: string,
-  newAuthTokenType?: string,
-  newRefreshToken?: string,
-  newAuthUserState?: T | null
-}
-
-export interface AuthKitSetState<T> {
-  auth?: {
-    token: string,
-    type: string
-  } | null,
-  refresh?: string | null,
-  userState?: T
-}
-
-interface AuthKitStateInterfaceTrue<T> {
-  auth: {
-    token: string,
-    type: string,
-    expiresAt: Date
-  },
-  refresh: {
-    token: string,
-    expiresAt: Date
-  } | null,
-  userState: T | null,
-  isSignIn: boolean,
-  isUsingRefreshToken: boolean,
-}
-
-interface AuthKitStateInterfaceNoAuthOnlyRefresh {
-  auth: null,
-  refresh: {
-    token: string,
-    expiresAt: Date
-  },
-  userState: null,
-  isSignIn: boolean,
-  isUsingRefreshToken: boolean,
-}
-
-
-interface AuthKitStateInterfaceFalse {
-  auth: null,
-  refresh: null,
-  userState: null,
-  isSignIn: boolean,
-  isUsingRefreshToken: boolean,
-}
-
-export type AuthKitStateInterface<T> = AuthKitStateInterfaceTrue<T> | AuthKitStateInterfaceFalse | AuthKitStateInterfaceNoAuthOnlyRefresh
-
-
 /**
- * SignIn function param
- * Used by: useSignIn and withSignIn
+ * Parameters used by the Sign In function
  */
 export type signInFunctionParams<T> = SignInActionPayload<T>
 
 /**
- * Context values type
+ * Auth State Object
+ * 1. when the user is authenticated
+ * 2. Refresh token can be there or not
+ * 3. User state can be there or not
  */
-// export interface AuthContextInterface<T> {
-//   authState: AuthKitStateInterface<T>
-//   dispatch: React.Dispatch<AuthActions>
-// }
+interface AuthKitStateInterfaceTrue<T> {
+  
+  /**
+   * Authentication part of the State
+   */
+  auth: {
+    
+    /**
+     * Auth access token
+     */
+    token: string,
 
-/**
- * Refresh Token Callback Response
- */
-export type RefreshTokenCallbackResponse<T> = {
-  isSuccess: boolean,
-  newAuthToken: string,
-  newRefreshToken?: string,
-  newAuthUserState?: T | null,
-};
+    /**
+     * Type of the Token
+     * 
+     * @example
+     * Bearer
+     */
+    type: string,
+    
+    /**
+     * Expiry time of the auth token
+     */
+    expiresAt: Date
+  },
 
-/**
- * Refresh Token types
- */
-// Callback function
+  /**
+   * Auth Refresh token part
+   */
+  refresh: {
 
-export type refreshTokenCallback<T> = (param: {
-  authToken?: string,
-  refreshToken?: string,
-  authUserState: T | null,
-}) => Promise<RefreshTokenCallbackResponse<T>>
+    /**
+     * Refresh token
+     */
+    token: string,
 
-// createRefresh function parameter
-export interface createRefreshParamInterface<T> {
-  interval: number,
-  refreshApiCallback: refreshTokenCallback<T>
+    /**
+     * Expiry time of the Refresh token
+     */
+    expiresAt: Date
+  } | null,
+
+  /**
+   * User state
+   */
+  userState: T | null,
+  
+  /**
+   * Boolean value to know if the user is signed in or not
+   */
+  isSignIn: boolean,
+
+  /**
+   * If the current system is using refresh token or not
+   */
+  isUsingRefreshToken: boolean,
 }
 
-export type refreshFunctionType<T> = (param: createRefreshParamInterface<T>)
-  => createRefreshParamInterface<T>
+/**
+ * Auth State Object
+ * 1. when the user is not authenticated
+ * 2. Refresh token is there
+ * 3. User state is null
+ */
+interface AuthKitStateInterfaceNoAuthOnlyRefresh {
+  
+  /**
+   * Authentication part of the State
+   */
+  auth: null,
+
+  /**
+   * Auth Refresh token part
+   */
+  refresh: {
+
+    /**
+     * Refresh token
+     */
+    token: string,
+
+    /**
+     * Expiry time of the Refresh token
+     */
+    expiresAt: Date
+  },
+
+  /**
+   * User State
+   */
+  userState: null,
+
+  /**
+   * Boolean value to know if the user is signed in or not
+   */
+  isSignIn: boolean,
+
+  /**
+   * If the current system is using refresh token or not
+   */
+  isUsingRefreshToken: boolean,
+}
+
+/**
+ * Auth State Object
+ * 1. when the user is not authenticated
+ * 2. Refresh token is null
+ * 3. User state is null
+ */
+interface AuthKitStateInterfaceFalse {
+
+  /**
+   * Authentication part of the State, which is null
+   */
+  auth: null,
+  
+  /**
+   * Refresh token part, which is null
+   */
+  refresh: null,
+  
+  /**
+   * User state, which is null
+   */
+  userState: null,
+  
+  /**
+   * Is Sign In, which is false
+   */
+  isSignIn: boolean,
+  
+  /**
+   * If the current system is using refresh token or not
+   */
+  isUsingRefreshToken: boolean,
+}
+
+/**
+ * Auth State Object
+ */
+export type AuthKitStateInterface<T> = AuthKitStateInterfaceTrue<T> | AuthKitStateInterfaceFalse | AuthKitStateInterfaceNoAuthOnlyRefresh
